--- interface/intarface.cpp
@@ -1,10 +1,0 @@
-//
-// Created by Тимофей Савинич on 19.03.24.
-//
-
-#include "intarface.h"
-
-void interface()
-{
-
-}
