--- interface/intarface.h
@@ -1,10 +1,0 @@
-//
-// Created by Тимофей Савинич on 19.03.24.
-//
-
-#ifndef COURSE_WORK_INTARFACE_H
-#define COURSE_WORK_INTARFACE_H
-
-void interface();
-
-#endif //COURSE_WORK_INTARFACE_H
