+++ interface/interface.h
@@ -1,0 +1,10 @@
+//
+// Created by Тимофей Савинич on 19.03.24.
+//
+
+#ifndef COURSE_WORK_INTERFACE_H
+#define COURSE_WORK_INTERFACE_H
+
+void interface();
+
+#endif //COURSE_WORK_INTERFACE_H
